functions/webhook/requirements.txt
functions-framework==3.4.0
google-cloud-pubsub==2.18.3
flask==2.3.3